feat: split logging_config.py into separate logging and output modules
- Create new output_formatting.py module containing ConsoleOutput class and Rich dependencies
- Update logging_config.py to focus only on logging functionality (structlog, file handlers)
- Update all Python scripts (init.py, swman.py, pkgstatus.py) to import from both modules
- Maintain full backward compatibility for all functionality
- Follow single responsibility principle with clear module separation

Addresses issue #5 - logging config file had mixed responsibilities for both
logging configuration and console output formatting.
